Create TestA
This test is run to determine if this configuration of the 6 fans will cause the entire UFO to rotate with the direction of the EDF shafts. All 6 fans are run at an arbitrary low thrust value while the apparatus is suspended by fishing wire. If the fan fails the test and begins to rotate with the EDF shafts, another configuration is needed.