Create TestC
http://robots.dacloughb.com/project-2/esc-calibration-programming/

This experiment evaluates how the UFO will change directions - by lowering the thrust value of the left fans, and holding the other 4 fans at their 'hover value', the entire UFO should tilt to the left and cause the apparatus to swing to the left. If this test fails, the group will consider raising the value of the right two EDFs in order to help the UFO turn.